Adjusted the layout
I changed the layout on the part of projects on the index.html so that the projects.html has more details concerning the projects and also the index.html will look less congested. All we are left with is adding the actual content and other pics. I talked to the president and he said the website looks good and he will send a document of the content they would want us to add. I'll deal with that.